dcp-815 rename terra to terra.experiment
